Move LD facilities into /ecom
The Pos SDK does not use these; in fact we dropped LD from the Pos repo
a long time ago
